mdification du type du montant en Double

+++ src/main/java/com/group3/persobudgetmanager/models/Budget.java
@@ -26,7 +26,7 @@ public class Budget {
 
     @Column(name = "montant")
     @NotNull(message = "")
-    private Long amount; // le montant du budget
+    private Double amount; // le montant du budget
 
     @Column(name = "montant_alert")
     @NotNull(message = "")
